state: Move indices to checksums.
* For space optimization reasons, we were using uint64 indices in place
of the real checksums. This commit moves us back to using checksum
indices. This is indeed bad but it's in preparation for moving the state
handling to disk rather than memory, where we don't really care about
space but we'd like to get rid of the complexity of mapping checksums.

* Warning: this breaks the ondisk format of repository so be sure to
trash your repo and caches before using it.