importer/fs: add prefixes also for the "original" path
We backup the "real" directory, i.e. if the argument points to a
symlink we backup what it points to.  Yet, the backup root directory
remains the symlinked dir and this breaks the navigation in the
backup, and also the user expectations.

For clarity: we only follow backup on the first argument, i.e. the
target of the backup, not in other symlinks found during the scan.
This only applies to that case.